# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for Recally project.",
        "title": "Recally API",
        "termsOfService": "https://recally.vaayne.com/terms/",
        "contact": {
            "name": "Vaayne",
            "url": "https://vaayne.com",
            "email": "recally@vaayne.com"
        },
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/api/v1",
    "paths": {
...
        "/files/presigned-urls": {
            "get": {
                "description": "Get presigned URLs for uploading or downloading files from S3",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get presigned URLs for file operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the file",
                        "name": "fileName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MIME type of the file",
                        "name": "fileType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "put",
                            "get"
                        ],
                        "type": "string",
                        "description": "Action to perform (put or get)",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Expiration time in seconds (max 604800)",
                        "name": "expiration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.getPresignedURLsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Get a redirect to the file's presigned URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Redirect to file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to file URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Object not found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a file by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete a file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpserver.JSONResult": {
            "description": "JSONResult represents the structure of the JSON response.",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is an integer value that represents the HTTP status code.",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "Data is an interface value that represents the data of the response."
                },
                "error": {
                    "description": "Error is an error value that represents the error of the response."
                },
                "message": {
                    "description": "Message is a string value that represents the message of the response.",
                    "type": "string",
                    "example": "OK"
                },
                "success": {
                    "description": "Success is a boolean value that indicates whether the request was successful.",
                    "type": "boolean"
                }
            }
        },
...
        "httpserver.getPresignedURLsResponse": {
            "type": "object",
            "properties": {
                "object_key": {
                    "type": "string"
                },
                "presigned_url": {
                    "type": "string"
                },
                "public_url": {
                    "type": "string"
                }
            }
        },
...
    }
}